make build.yml reproducible
`build.yml` is currently biased to succeed every time on this repository. Reasoning is that it will always checkout the actions' repository (which essentially transfers the actions' files to the "working directory", not the case if this action is being used externally), and it assumes that the action is run locally.

This commit breaks the build file into two parts: one that performs the current action as-is, and a new job which attempts to simulate usage from an external repository. In this job, no initial checkout is made and an immediate attempt at connecting to a ZeroTier network is made.