action.yml: force use external syntax for `uses`
This forces the use of the syntax for `uses` calling a specified subfolder of a public repository at a given ref. This is an attempt at moving away from the use of `./` as a path, which forces GitHub Actions to locate the `util/post/action.yml` of the *working directory* and not of the action path. This exists since `uses` cannot take expressions; this would be an easy fix if `./` can be replaced with `${{ github.action_path }}/`. Alas, this is not possible, and this hacky workaround exists instead.