build.yml: only run test job on main
This forces the `test` job to only run if the workflow was triggered from the `main` branch. It wouldn't make sense to check if `main` works when using a different branch or tag, so this avoids doing that entirely.

`test` cannot automatically use the ref that triggered this workflow because `uses` does not take in expressions (and therefore cannot use `github.ref_name`).